feature : can delete a customer but error for parameter iduser in deleteUser(), need to solve it

+++ app/Http/Controllers/UserController.php
@@ -109,6 +109,7 @@ class UserController extends Controller
 
     public function deleteUser(int $iduser)
     {
+        if (empty($iduser)) return Redirect::route('dashboard/customers');
         $user = User::find($iduser);
         $user->delete();
         return Redirect::route('dashboard');
